+++ app/assets/javascripts/spina/admin/controllers/datetime_picker_controller.js
@@ -1,0 +1,7 @@
+(() => {
+  const application = Stimulus.Application.start()
+
+  application.register("datetime-picker", class extends StimulusFlatpickr {
+
+  })
+})()