+++ app/models/spina/parts/datetime.rb
@@ -1,0 +1,9 @@
+module Spina
+  module Parts
+    class Datetime < Base
+      attr_json :content, :datetime, default: ""
+
+      attr_accessor :options
+    end
+  end
+end